backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
Pillow==10.4.0
numpy==2.1.3
opencv-python==4.10.0.84
scikit-image==0.24.0
python-multipart==0.0.20
albumentations==1.4.18
tqdm==4.66.5
openai>=1.0.0

bioimageio.core>=0.6.8
